voicebot-backend/requirements.txt
flask
boto3
python-dotenv
requests